+++ doc/changelog.rst
@@ -23,6 +23,7 @@ Core
 - Add ``DiscountModule`` interface for pricing
 - Simplify ``PricingContext`` and require shop and customer for it
 - Add ``get_price_info`` etc. functions to ``shoop.core.pricing``
+- Add ``get_suppliable_products`` to ``shoop.core.models.Supplier``
 - Add "codes" API to OrderSource and BaseBasket
 - Reword doc/provides.rst
 
@@ -45,6 +46,7 @@ Front
 - Check product quantity already in basket while adding
 - Process given coupon codes in basket
 - Process discounts from new campaign engine
+- Add ``get_visible_products`` template helper
 
 Xtheme
 ~~~~~~
@@ -55,6 +57,7 @@ Classic Gray Theme
 
 - Show error messages while adding products to basket
 - Add Coupon use possibility to basket page
+- Add option to only show orderable products to highlights plugin
 
 Simple Supplier
 ~~~~~~~~~~~~~~~